add beta site and introductory page
The beta.cpantesters.org site is not the main web app, but right now it
looks like it. It also has the cool chart thing I did, but it's not
linked from anywhere. So to make the cool projects we're doing more
accessible, here's an introduction page that runs when the "mode" is
"beta".