add schema and support commands to web app
We're using DBIx::Class::Schema::Versioned, which isn't very easy to
use, but is more comprehensible than DeploymentHandler. The command is
needed to prepare the schema upgrade files, install new schemas, and
upgrade existing schemas.

I should make something text-based like Mojo::*::Migrations...

We now have multiple DBIx::Class schemas in the website, and eventually
we'll have even more (one object per language). This is going to get
untenable fast, so we'll need to come up with a better way to handle
this...